+++ changelog.md
@@ -1,3 +1,7 @@
+# AvrixLoader v1.2.1
+
+- Fixed copying configs from the plugin archive
+
 # AvrixLoader v1.2.0
 
 - Added `getPlayerAccessLevel` method